+++ src/Exception/NonExistingFieldException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ArrowSphere\Entities\Exception;
+
+/**
+ * Class NonExistingFieldException
+ */
+class NonExistingFieldException extends EntitiesException
+{
+}